fixed the index file in Data and OtherStuff
